+++ db/migrate/20140331060454_create_expenses.rb
@@ -5,7 +5,7 @@ class CreateExpenses < ActiveRecord::Migration
       t.belongs_to :vendor
 
       t.string :name
-      t.integer :amount
+      t.float :amount
       t.text :description
       t.date :date
